+++ src/modules/other/index.ts
@@ -1,0 +1,1 @@
+export { default, otherController } from './controller';